+++ examples/run_remote_in_chrome_with_options_for_selenoid/test_todomvc.py
@@ -1,4 +1,4 @@
-import os
+import dotenv
 from selenium import webdriver
 from selene import browser, have
 
@@ -6,14 +6,24 @@ from selene import browser, have
 def test_complete_task():
     options = webdriver.ChromeOptions()
     options.browser_version = '100.0'
-    options.set_capability('screenResolution', '1920x1080x24')
-    options.set_capability('enableVNC', True)
-    options.set_capability('enableVideo', True)
+    options.set_capability(
+        'selenoid:options',
+        {
+            'screenResolution': '1920x1080x24',
+            'enableVNC': True,
+            'enableVideo': True,
+            'enableLog': True,
+        },
+    )
     browser.config.driver_options = options
+    project_config = dotenv.dotenv_values()
     browser.config.remote_url = (
-        f'https://{os.getenv("LOGIN")}:{os.getenv("PASSWORD")}@'
+        f'https://{project_config["LOGIN"]}:{project_config["PASSWORD"]}@'
         f'selenoid.autotests.cloud/wd/hub'
     )
+    # To speed tests a bit
+    # by not checking if driver is alive before each action
+    browser.config.rebuild_dead_driver = False
 
     browser.open('http://todomvc.com/examples/emberjs/')
     browser.should(have.title_containing('TodoMVC'))