+++ tests/integration/conftest.py
@@ -1,3 +1,7 @@
+import dataclasses
+import os
+
+import dotenv
 import pytest
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service as ChromeService
@@ -27,6 +31,74 @@ def chrome_driver(request):
     chrome_driver.quit()
 
 
+@pytest.fixture(scope='function')
+def a_remote_browser():
+    options = webdriver.ChromeOptions()
+    options.browser_version = '100.0'
+    options.set_capability(
+        'selenoid:options',
+        {
+            'screenResolution': '1920x1080x24',
+            'enableVNC': True,
+            'enableVideo': True,
+            'enableLog': True,
+        },
+    )
+    project_config = dotenv.dotenv_values()
+    browser_ = Browser(
+        Config(
+            driver_options=options,
+            remote_url=(
+                f'https://{project_config["LOGIN"]}:{project_config["PASSWORD"]}@'
+                f'selenoid.autotests.cloud/wd/hub'
+            ),
+        )
+    )
+
+    yield browser_
+
+    browser_.quit()
+
+
+@pytest.fixture(scope='module')
+def the_module_remote_browser():
+    options = webdriver.ChromeOptions()
+    options.browser_version = '100.0'
+    options.set_capability(
+        'selenoid:options',
+        {
+            'screenResolution': '1920x1080x24',
+            'enableVNC': True,
+            'enableVideo': True,
+            'enableLog': True,
+        },
+    )
+
+    @dataclasses.dataclass
+    class ProjectConfig:
+        LOGIN: str = dataclasses.field(
+            default_factory=lambda: os.getenv('LOGIN', None)
+        )
+        PASSWORD: str = dataclasses.field(
+            default_factory=lambda: os.getenv('PASSWORD', None)
+        )
+
+    project_config = ProjectConfig(**dotenv.dotenv_values())
+    browser_ = Browser(
+        Config(
+            driver_options=options,
+            remote_url=(
+                f'https://{project_config.LOGIN}:{project_config.PASSWORD}@'
+                f'selenoid.autotests.cloud/wd/hub'
+            ),
+        )
+    )
+
+    yield browser_
+
+    browser_.quit()
+
+
 @pytest.fixture(scope='session')
 def firefox_driver(request):
     gecko_driver = webdriver.Firefox(
@@ -36,6 +108,9 @@ def firefox_driver(request):
     gecko_driver.quit()
 
 
+# TODO: in fact its a_browser_with_session_driver,
+#       where a_ means a new one with fresh config, but old same driver
+#       should we rename it?
 @pytest.fixture(scope='function')
 def session_browser(chrome_driver):
     yield Browser(Config(driver=chrome_driver))